More breaking up of inputModifiers
